Crawl a url and return urls & assets in json format
The Crawler class crawls a given url and returns a json report of the site's urls and assets. It ignored external urls and urls on subdomains; and it also ignores assets from outside the same domain.

The Crawler is run from the terminal like so:
`./bin/crawler http://www.livestax.com`
(assuming http://www.livestax.com is the url you wish to crawl)