Fix code smell in appending page crawl output to main output
I had an issue with (what I thought was) a default `0` being appended to the @output. It turns out that it wasn't a default value at all, but a by-product of my appending results to the output in two places. Now the output is appended ONCE, and this issue goes away.