Add another test for actual crawling
The existing single successful test only tested the output format. This additional test also tests the crawlers ability to crawl, through 3 very simple pages.

This test also exposed a small error where the `assets: []` array was appearing even if it was empty. I am now not showing the assets array in the output if it is empty.